+++ Packages/Syntax/Sources/SyntaxParsers/Queries/JavaScript/highlights.scm
@@ -91,10 +91,6 @@
   left: (identifier) @commands
   right: [(function_expression) (arrow_function)])
 
-; built-in functions
-((identifier) @commands
- (#eq? @commands "require"))
-
 
 ; MARK: Types
 ; ----------------------------