+++ Packages/Syntax/Sources/SyntaxParsers/Queries/TypeScript/highlights.scm
@@ -108,10 +108,6 @@
   left: (identifier) @commands
   right: [(function_expression) (arrow_function)])
 
-; built-in functions
-((identifier) @commands
- (#eq? @commands "require"))
-
 
 ; MARK: Types
 ; ----------------------------